- Fixed build error when -fpermissive compiler flag is set.
git-svn-id: svn://svn.code.sf.net/p/wxsf/code/trunk@412 1a166092-e534-0410-9de8-debdbcc2e763

+++ src/ShapeCanvas.cpp
@@ -3309,7 +3309,7 @@ void wxSFShapeCanvas::RestorePrevPositions()
 
 void wxSFShapeCanvas::StorePrevPosition(const wxSFShapeBase* shape)
 {
-	m_mapPrevPositions[ (int)shape ] = new wxRealPoint( shape->GetRelativePosition() );
+	m_mapPrevPositions[ (long)shape ] = new wxRealPoint( shape->GetRelativePosition() );
 }
 
 /*#ifdef __WXMAC__
